feat(path): add function to find binary in PATH environment variable
This commit adds a new module `path` in the `tgs/src` directory. The module contains a function `find_binary` that looks for the specified binary command in the `PATH` environment variable. If the command is found, it returns the path to the binary file. If the command is not found, it returns an error.

feat(process): add function to run process with binary from PATH

This commit adds a new module `process` in the `tgs/src` directory. The module contains a function `run_process` that takes a vector of environment variables and a command as input. It uses the `find_binary` function from the `path` module to find the binary file corresponding to the command in the `PATH` environment variable. Currently, the function does not perform any actual process execution, it just returns `Ok(())`.